// src/modulos/viajes/tipos/viaje.enums.ts
export enum TipoViaje {
  INSUMOS = 'INSUMOS',
  CAJAS = 'CAJAS',
}

export enum EstadoViaje {
  CREADO = 'CREADO',
  EN_RUTA = 'EN_RUTA',
  ENTREGADO = 'ENTREGADO',
  ANULADO = 'ANULADO',
}

export enum TipoDestino {
  HACIENDA = 'HACIENDA',
  BODEGA = 'BODEGA',
  CLIENTE = 'CLIENTE',
}

export enum InsumoTipo {
  CARTON = 'CARTON',
  FUNDA = 'FUNDA',
  CARTULINA = 'CARTULINA',
}